player movement works fine
