+++ Scripts/Characters/PlayerCharacter.cs
@@ -1,12 +1,12 @@
 using Godot;
 using System;
 
-public partial class PlayerCharacter : CharacterBody2D
+public partial class PlayerCharacter : Character
 {
 	// public const float Speed = 300.0f;
 	// public const float JumpVelocity = -400.0f;
 
-	public override void _PhysicsProcess(double delta)
+	/* public override void _PhysicsProcess(double delta)
 	{
 		Vector2 velocity = Velocity;
 
@@ -16,15 +16,13 @@ public partial class PlayerCharacter : CharacterBody2D
 		if (direction != Vector2.Zero)
 		{
 			velocity.X = direction.X * Speed;
-			velocity.Y = direction.Y * Speed;
 		}
 		else
 		{
 			velocity.X = Mathf.MoveToward(Velocity.X, 0, Speed);
-			velocity.Y = Mathf.MoveToward(Velocity.Y, 0, Speed);
 		}
 
 		Velocity = velocity;
 		MoveAndSlide();
-	}
+	} */
 }
